Config: Hotfix to prevent per-game settings from ie. GameINIs being stored to the global user configuration.

+++ Source/Core/Core/ConfigManager.cpp
@@ -7,6 +7,7 @@
 #include "Common/FileUtil.h"
 #include "Common/IniFile.h"
 #include "Core/ConfigManager.h"
+#include "Core/Core.h"
 #include "Core/HW/SI.h"
 #include "Core/PowerPC/PowerPC.h"
 #include "DiscIO/NANDContentLoader.h"
@@ -39,6 +40,13 @@ SConfig::~SConfig()
 
 void SConfig::SaveSettings()
 {
+	// TODO: The is a hotfix to prevent writing of temporary per-game settings
+	// (GameINI, Movie, Netplay, ...) to the global Dolphin configuration file.
+	// The Config logic should be rewritten instead so that per-game settings
+	// aren't stored in the same configuration as the actual user settings.
+	if (Core::IsRunning())
+		return;
+
 	NOTICE_LOG(BOOT, "Saving settings to %s", File::GetUserPath(F_DOLPHINCONFIG_IDX).c_str());
 	IniFile ini;
 	ini.Load(File::GetUserPath(F_DOLPHINCONFIG_IDX)); // load first to not kill unknown stuff
